+++ core/js-native/src/main/scala/cats/effect/unsafe/metrics/IORuntimeMetricsPlatform.scala
@@ -14,25 +14,6 @@
  * limitations under the License.
  */
 
-package cats.effect.metrics
+package cats.effect.unsafe.metrics
 
-/**
- * The runtime-specific metrics.
- */
-trait IORuntimeMetrics extends IORuntimeMetricsPlatform {
-
-  /**
-   * Returns starvation-specific metrics.
-   *
-   * @example
-   *   {{{
-   * val runtime: IORuntime = ???
-   * val maxDrift = runtime.metrics.cpuStarvation.clockDriftMax()
-   *   }}}
-   */
-  def cpuStarvation: CpuStarvationMetrics
-
-  private[effect] def cpuStarvationSampler: CpuStarvationSampler
-}
-
-object IORuntimeMetrics extends IORuntimeMetricsCompanionPlatform
+private[metrics] trait IORuntimeMetricsPlatform { this: IORuntimeMetrics => }
