+++ core/jvm/src/main/scala/cats/effect/unsafe/metrics/IORuntimeMetricsPlatform.scala
@@ -14,17 +14,6 @@
  * limitations under the License.
  */
 
-package cats.effect.metrics
+package cats.effect.unsafe.metrics
 
-private[metrics] abstract class IORuntimeMetricsCompanionPlatform {
-  this: IORuntimeMetrics.type =>
-
-  private[effect] def apply(): IORuntimeMetrics =
-    new IORuntimeMetrics {
-      private[effect] val cpuStarvationSampler: CpuStarvationSampler =
-        CpuStarvationSampler()
-
-      val cpuStarvation: CpuStarvationMetrics =
-        CpuStarvationMetrics(cpuStarvationSampler)
-    }
-}
+private[metrics] trait IORuntimeMetricsPlatform { this: IORuntimeMetrics => }
