+++ core/shared/src/main/scala/cats/effect/CpuStarvationCheck.scala
@@ -16,9 +16,10 @@
 
 package cats.effect
 
-import cats.effect.metrics.{CpuStarvationSampler, CpuStarvationWarningMetrics}
+import cats.effect.metrics.CpuStarvationWarningMetrics
 import cats.effect.std.Console
 import cats.effect.unsafe.IORuntimeConfig
+import cats.effect.unsafe.metrics.CpuStarvationSampler
 import cats.syntax.all._
 
 import scala.concurrent.duration.{Duration, FiniteDuration}
